+++ CHANGELOG.md
@@ -8,8 +8,9 @@ All notable changes to this project will be documented in this file.
 - Fixed "taken IsoGenerator" line in *_map.txt log file.
 
 ### Added
-- Added profession to player log.
+- Added profession and skill points to player log in stats section.
 - Added coordinates to player log.
+- Added health level and infected information to player log.
 
 ## [v0.3.0] - 2019-12-01
 ### Added